CRM-1610: Create migrations for incompatible initial migrations - zendesk bundle

+++ Migrations/Schema/v1_0/OroCRMZendeskBundle.php
@@ -80,7 +80,7 @@ class OroCRMZendeskBundle implements Migration
         $table->addColumn('priority_name', 'string', ['notnull' => false, 'length' => 16]);
         $table->addColumn('requester_id', 'integer', ['notnull' => false]);
         $table->addColumn('case_id', 'integer', ['notnull' => false]);
-        $table->addColumn('channel_id', 'smallint', ['notnull' => false]);
+        $table->addColumn('channel_id', 'integer', ['notnull' => false]);
         $table->addColumn('type_name', 'string', ['notnull' => false, 'length' => 16]);
         $table->addColumn('origin_id', 'bigint', ['notnull' => false]);
         $table->addColumn('url', 'string', ['notnull' => false, 'length' => 255]);
@@ -174,7 +174,7 @@ class OroCRMZendeskBundle implements Migration
         /** Generate table orocrm_zd_user **/
         $table = $schema->createTable('orocrm_zd_user');
         $table->addColumn('id', 'integer', ['autoincrement' => true]);
-        $table->addColumn('channel_id', 'smallint', ['notnull' => false]);
+        $table->addColumn('channel_id', 'integer', ['notnull' => false]);
         $table->addColumn('related_contact_id', 'integer', ['notnull' => false]);
         $table->addColumn('related_user_id', 'integer', ['notnull' => false]);
         $table->addColumn('role_name', 'string', ['notnull' => false, 'length' => 16]);
@@ -330,7 +330,7 @@ class OroCRMZendeskBundle implements Migration
         /** Generate table orocrm_zd_comment **/
         $table = $schema->createTable('orocrm_zd_comment');
         $table->addColumn('id', 'integer', ['autoincrement' => true]);
-        $table->addColumn('channel_id', 'smallint', ['notnull' => false]);
+        $table->addColumn('channel_id', 'integer', ['notnull' => false]);
         $table->addColumn('related_comment_id', 'integer', ['notnull' => false]);
         $table->addColumn('author_id', 'integer', ['notnull' => false]);
         $table->addColumn('ticket_id', 'integer', ['notnull' => false]);
